refactor: modularize page component types
