+++ packages/types/src/page/base.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { localeSchema, type Translated } from "../Product";
+import type { Translated } from "../Product";
 
 export type PageStatus = "draft" | "published";
 
@@ -102,299 +102,7 @@ export interface PageComponentBase {
   [key: string]: unknown;
 }
 
-export interface HeaderComponent extends PageComponentBase {
-  type: "Header";
-  nav?: { label: string; url: string }[];
-  logo?: string;
-}
-
-export interface FooterComponent extends PageComponentBase {
-  type: "Footer";
-  links?: { label: string; url: string }[];
-  logo?: string;
-}
-
-export interface SocialLinksComponent extends PageComponentBase {
-  type: "SocialLinks";
-  facebook?: string;
-  instagram?: string;
-  x?: string;
-  youtube?: string;
-  linkedin?: string;
-}
-
-export interface SocialFeedComponent extends PageComponentBase {
-  type: "SocialFeed";
-  platform?: "twitter" | "instagram";
-  account?: string;
-  hashtag?: string;
-}
-
-export interface SocialProofComponent extends PageComponentBase {
-  type: "SocialProof";
-  source?: string;
-  frequency?: number;
-}
-
-export interface AnnouncementBarComponent extends PageComponentBase {
-  type: "AnnouncementBar";
-  text?: string;
-  link?: string;
-  closable?: boolean;
-}
-
-export interface HeroBannerComponent extends PageComponentBase {
-  type: "HeroBanner";
-  slides?: { src: string; alt?: string; headlineKey: string; ctaKey: string }[];
-}
-
-export interface ValuePropsComponent extends PageComponentBase {
-  type: "ValueProps";
-  items?: { icon: string; title: string; desc: string }[];
-}
-
-/** Carousel of customer reviews. `minItems`/`maxItems` limit visible reviews */
-export interface ReviewsCarouselComponent extends PageComponentBase {
-  type: "ReviewsCarousel";
-  reviews?: { nameKey: string; quoteKey: string }[];
-}
-
-/** Grid of products; `minItems`/`maxItems` clamp the responsive product count */
-export interface ProductGridComponent extends PageComponentBase {
-  type: "ProductGrid";
-  skus?: string[];
-  collectionId?: string;
-  mode?: "collection" | "manual";
-  /** Enable product quick view modal */
-  quickView?: boolean;
-}
-
-/** Carousel of products; `minItems`/`maxItems` clamp visible products */
-export interface ProductCarouselComponent extends PageComponentBase {
-  type: "ProductCarousel";
-  skus?: string[];
-  collectionId?: string;
-  mode?: "collection" | "manual";
-  /** Enable product quick view modal */
-  quickView?: boolean;
-}
-
-/** Grid of collections; `minItems`/`maxItems` clamp visible collections */
-export interface CollectionListComponent extends PageComponentBase {
-  type: "CollectionList";
-  collections?: { id: string; title: string; image: string }[];
-  desktopItems?: number;
-  tabletItems?: number;
-  mobileItems?: number;
-}
-
-/** Carousel of recommended products fetched from an API. `minItems`/`maxItems` limit visible products */
-export interface RecommendationCarouselComponent extends PageComponentBase {
-  type: "RecommendationCarousel";
-  endpoint: string;
-}
-
-export interface GalleryComponent extends PageComponentBase {
-  type: "Gallery";
-  images?: { src: string; alt?: string }[];
-}
-
-export interface LookbookComponent extends PageComponentBase {
-  type: "Lookbook";
-  src?: string;
-  alt?: string;
-  hotspots?: { sku?: string; x: number; y: number }[];
-}
-
-export interface ImageSliderComponent extends PageComponentBase {
-  type: "ImageSlider";
-  slides?: { src: string; alt?: string; caption?: string }[];
-  minItems?: number;
-  maxItems?: number;
-}
-
-export interface ContactFormComponent extends PageComponentBase {
-  type: "ContactForm";
-  action?: string;
-  method?: string;
-}
-
-export interface NewsletterSignupComponent extends PageComponentBase {
-  type: "NewsletterSignup";
-  text?: string;
-  action?: string;
-  placeholder?: string;
-  submitLabel?: string;
-}
-
-export interface SearchBarComponent extends PageComponentBase {
-  type: "SearchBar";
-  placeholder?: string;
-  limit?: number;
-}
-
-export interface ContactFormWithMapComponent extends PageComponentBase {
-  type: "ContactFormWithMap";
-  mapSrc?: string;
-}
-
-export interface MapBlockComponent extends PageComponentBase {
-  type: "MapBlock";
-  lat?: number;
-  lng?: number;
-  zoom?: number;
-}
-
-export interface StoreLocatorBlockComponent extends PageComponentBase {
-  type: "StoreLocatorBlock";
-  locations?: { lat?: number; lng?: number; label?: string }[];
-  zoom?: number;
-}
-
-export interface VideoBlockComponent extends PageComponentBase {
-  type: "VideoBlock";
-  src?: string;
-  autoplay?: boolean;
-}
-
-export interface FAQBlockComponent extends PageComponentBase {
-  type: "FAQBlock";
-  items?: { question: string; answer: string }[];
-}
-
-export interface CountdownTimerComponent extends PageComponentBase {
-  type: "CountdownTimer";
-  targetDate?: string;
-  timezone?: string;
-  completionText?: string;
-  styles?: string;
-}
-
-export interface ImageComponent extends PageComponentBase {
-  type: "Image";
-  src?: string;
-  alt?: string;
-}
-
-export interface BlogListingComponent extends PageComponentBase {
-  type: "BlogListing";
-  posts?: { title: string; excerpt?: string; url?: string }[];
-}
-
-/** Slider of testimonials. `minItems`/`maxItems` limit visible testimonials */
-export interface TestimonialSliderComponent extends PageComponentBase {
-  type: "TestimonialSlider";
-  testimonials?: { quote: string; name?: string }[];
-}
-
-export interface TestimonialsComponent extends PageComponentBase {
-  type: "Testimonials";
-  testimonials?: { quote: string; name?: string }[];
-}
-
-export interface PricingTableComponent extends PageComponentBase {
-  type: "PricingTable";
-  plans?: {
-    title: string;
-    price: string;
-    features: string[];
-    ctaLabel: string;
-    ctaHref: string;
-    featured?: boolean;
-  }[];
-}
-
-export interface GiftCardBlockComponent extends PageComponentBase {
-  type: "GiftCardBlock";
-  denominations?: number[];
-  description?: string;
-}
-
-export interface PopupModalComponent extends PageComponentBase {
-  type: "PopupModal";
-  trigger?: "load" | "delay" | "exit";
-  delay?: number;
-  content?: string;
-}
-
-export interface TextComponent extends PageComponentBase {
-  type: "Text";
-  text?: string;
-}
-
-export interface CustomHtmlComponent extends PageComponentBase {
-  type: "CustomHtml";
-  html?: string;
-}
-
-export interface ButtonComponent extends PageComponentBase {
-  type: "Button";
-  label?: string;
-  href?: string;
-  variant?: "default" | "outline" | "ghost" | "destructive";
-  size?: "sm" | "md" | "lg";
-}
-
-export interface SectionComponent extends PageComponentBase {
-  type: "Section";
-  children?: PageComponent[];
-}
-
-export interface MultiColumnComponent extends PageComponentBase {
-  type: "MultiColumn";
-  columns?: number;
-  gap?: string;
-  children?: PageComponent[];
-}
-
-export interface TabsComponent extends PageComponentBase {
-  type: "Tabs";
-  labels?: string[];
-  active?: number;
-  children?: PageComponent[];
-}
-
-export type PageComponent =
-  | AnnouncementBarComponent
-  | HeroBannerComponent
-  | ValuePropsComponent
-  | ReviewsCarouselComponent
-  | ProductGridComponent
-  | ProductCarouselComponent
-  | CollectionListComponent
-  | RecommendationCarouselComponent
-  | GalleryComponent
-  | LookbookComponent
-  | ImageSliderComponent
-  | ContactFormComponent
-  | NewsletterSignupComponent
-  | SearchBarComponent
-  | ContactFormWithMapComponent
-  | MapBlockComponent
-  | StoreLocatorBlockComponent
-  | VideoBlockComponent
-  | FAQBlockComponent
-  | CountdownTimerComponent
-  | BlogListingComponent
-  | TestimonialsComponent
-  | PricingTableComponent
-  | TestimonialSliderComponent
-  | GiftCardBlockComponent
-  | PopupModalComponent
-  | ImageComponent
-  | TextComponent
-  | CustomHtmlComponent
-  | ButtonComponent
-  | HeaderComponent
-  | FooterComponent
-  | SocialLinksComponent
-  | SocialFeedComponent
-  | SocialProofComponent
-  | SectionComponent
-  | MultiColumnComponent
-  | TabsComponent;
-
-const baseComponentSchema = z
+export const baseComponentSchema = z
   .object({
     id: z.string(),
     width: z.string().optional(),
@@ -424,364 +132,3 @@ const baseComponentSchema = z
   })
   .passthrough();
 
-const heroBannerComponentSchema = baseComponentSchema.extend({
-  type: z.literal("HeroBanner"),
-  slides: z
-    .array(
-      z.object({
-        src: z.string(),
-        alt: z.string().optional(),
-        headlineKey: z.string(),
-        ctaKey: z.string(),
-      })
-    )
-    .optional(),
-});
-
-const announcementBarComponentSchema = baseComponentSchema.extend({
-  type: z.literal("AnnouncementBar"),
-  text: z.string().optional(),
-  link: z.string().optional(),
-  closable: z.boolean().optional(),
-});
-
-const valuePropsComponentSchema = baseComponentSchema.extend({
-  type: z.literal("ValueProps"),
-  items: z
-    .array(z.object({ icon: z.string(), title: z.string(), desc: z.string() }))
-    .optional(),
-});
-
-const reviewsCarouselComponentSchema = baseComponentSchema.extend({
-  type: z.literal("ReviewsCarousel"),
-  reviews: z
-    .array(z.object({ nameKey: z.string(), quoteKey: z.string() }))
-    .optional(),
-});
-
-const productGridComponentSchema = baseComponentSchema.extend({
-  type: z.literal("ProductGrid"),
-});
-
-const productCarouselComponentSchema = baseComponentSchema.extend({
-  type: z.literal("ProductCarousel"),
-});
-
-const recommendationCarouselComponentSchema = baseComponentSchema.extend({
-  type: z.literal("RecommendationCarousel"),
-  endpoint: z.string(),
-});
-
-const galleryComponentSchema = baseComponentSchema.extend({
-  type: z.literal("Gallery"),
-  images: z
-    .array(z.object({ src: z.string(), alt: z.string().optional() }))
-    .optional(),
-});
-
-const lookbookComponentSchema = baseComponentSchema.extend({
-  type: z.literal("Lookbook"),
-  src: z.string().optional(),
-  alt: z.string().optional(),
-  hotspots: z
-    .array(
-      z.object({
-        sku: z.string().optional(),
-        x: z.number(),
-        y: z.number(),
-      })
-    )
-    .optional(),
-});
-
-const contactFormComponentSchema = baseComponentSchema.extend({
-  type: z.literal("ContactForm"),
-  action: z.string().optional(),
-  method: z.string().optional(),
-});
-
-const newsletterSignupComponentSchema = baseComponentSchema.extend({
-  type: z.literal("NewsletterSignup"),
-  text: z.string().optional(),
-  action: z.string().optional(),
-  placeholder: z.string().optional(),
-  submitLabel: z.string().optional(),
-});
-
-const searchBarComponentSchema = baseComponentSchema.extend({
-  type: z.literal("SearchBar"),
-  placeholder: z.string().optional(),
-  limit: z.number().optional(),
-});
-
-const contactFormWithMapComponentSchema = baseComponentSchema.extend({
-  type: z.literal("ContactFormWithMap"),
-  mapSrc: z.string().optional(),
-});
-
-const mapBlockComponentSchema = baseComponentSchema.extend({
-  type: z.literal("MapBlock"),
-  lat: z.number().optional(),
-  lng: z.number().optional(),
-  zoom: z.number().optional(),
-});
-
-const storeLocatorBlockComponentSchema = baseComponentSchema.extend({
-  type: z.literal("StoreLocatorBlock"),
-  locations: z
-    .array(
-      z.object({
-        lat: z.number().optional(),
-        lng: z.number().optional(),
-        label: z.string().optional(),
-      })
-    )
-    .optional(),
-  zoom: z.number().optional(),
-});
-
-const videoBlockComponentSchema = baseComponentSchema.extend({
-  type: z.literal("VideoBlock"),
-  src: z.string().optional(),
-  autoplay: z.boolean().optional(),
-});
-
-const faqBlockComponentSchema = baseComponentSchema.extend({
-  type: z.literal("FAQBlock"),
-  items: z
-    .array(
-      z.object({
-        question: z.string(),
-        answer: z.string(),
-      })
-    )
-    .optional(),
-});
-
-const countdownTimerComponentSchema = baseComponentSchema.extend({
-  type: z.literal("CountdownTimer"),
-  targetDate: z.string().optional(),
-  timezone: z.string().optional(),
-  completionText: z.string().optional(),
-  styles: z.string().optional(),
-});
-
-const headerComponentSchema = baseComponentSchema.extend({
-  type: z.literal("Header"),
-  nav: z
-    .array(z.object({ label: z.string(), url: z.string() }))
-    .optional(),
-  logo: z.string().optional(),
-});
-
-const footerComponentSchema = baseComponentSchema.extend({
-  type: z.literal("Footer"),
-  links: z
-    .array(z.object({ label: z.string(), url: z.string() }))
-    .optional(),
-  logo: z.string().optional(),
-});
-
-const socialLinksComponentSchema = baseComponentSchema.extend({
-  type: z.literal("SocialLinks"),
-  facebook: z.string().optional(),
-  instagram: z.string().optional(),
-  x: z.string().optional(),
-  youtube: z.string().optional(),
-  linkedin: z.string().optional(),
-});
-
-const socialFeedComponentSchema = baseComponentSchema.extend({
-  type: z.literal("SocialFeed"),
-  platform: z.enum(["twitter", "instagram"]).optional(),
-  account: z.string().optional(),
-  hashtag: z.string().optional(),
-});
-
-const socialProofComponentSchema = baseComponentSchema.extend({
-  type: z.literal("SocialProof"),
-  source: z.string().optional(),
-  frequency: z.number().optional(),
-});
-
-const blogListingComponentSchema = baseComponentSchema.extend({
-  type: z.literal("BlogListing"),
-  posts: z
-    .array(
-      z.object({
-        title: z.string(),
-        excerpt: z.string().optional(),
-        url: z.string().optional(),
-      })
-    )
-    .optional(),
-});
-
-const testimonialSliderComponentSchema = baseComponentSchema.extend({
-  type: z.literal("TestimonialSlider"),
-  testimonials: z
-    .array(z.object({ quote: z.string(), name: z.string().optional() }))
-    .optional(),
-});
-
-const testimonialsComponentSchema = baseComponentSchema.extend({
-  type: z.literal("Testimonials"),
-  testimonials: z
-    .array(z.object({ quote: z.string(), name: z.string().optional() }))
-    .optional(),
-});
-
-const pricingTableComponentSchema = baseComponentSchema.extend({
-  type: z.literal("PricingTable"),
-  plans: z
-    .array(
-      z.object({
-        title: z.string(),
-        price: z.string(),
-        features: z.array(z.string()),
-        ctaLabel: z.string(),
-        ctaHref: z.string(),
-        featured: z.boolean().optional(),
-      })
-    )
-    .optional(),
-});
-
-const giftCardBlockComponentSchema = baseComponentSchema.extend({
-  type: z.literal("GiftCardBlock"),
-  denominations: z.array(z.number()).optional(),
-  description: z.string().optional(),
-});
-
-const popupModalComponentSchema = baseComponentSchema.extend({
-  type: z.literal("PopupModal"),
-  trigger: z.enum(["load", "delay", "exit"]).optional(),
-  delay: z.number().int().min(0).optional(),
-  content: z.string().optional(),
-});
-
-const imageComponentSchema = baseComponentSchema.extend({
-  type: z.literal("Image"),
-  src: z.string().optional(),
-  alt: z.string().optional(),
-});
-
-const textComponentSchema = baseComponentSchema.extend({
-  type: z.literal("Text"),
-  text: z.string().optional(),
-});
-
-const customHtmlComponentSchema = baseComponentSchema.extend({
-  type: z.literal("CustomHtml"),
-  html: z.string().optional(),
-});
-
-const buttonComponentSchema = baseComponentSchema.extend({
-  type: z.literal("Button"),
-  label: z.string().optional(),
-  href: z.string().optional(),
-  variant: z
-    .enum(["default", "outline", "ghost", "destructive"])
-    .optional(),
-  size: z.enum(["sm", "md", "lg"]).optional(),
-});
-
-const sectionComponentSchema: z.ZodType<SectionComponent> =
-  baseComponentSchema.extend({
-    type: z.literal("Section"),
-    children: z.array(z.lazy(() => pageComponentSchema)).default([]),
-  });
-
-const multiColumnComponentSchema: z.ZodType<MultiColumnComponent> =
-  baseComponentSchema.extend({
-    type: z.literal("MultiColumn"),
-    columns: z.number().optional(),
-    gap: z.string().optional(),
-    children: z.array(z.lazy(() => pageComponentSchema)).default([]),
-  });
-
-const tabsComponentSchema: z.ZodType<TabsComponent> =
-  baseComponentSchema.extend({
-    type: z.literal("Tabs"),
-    labels: z.array(z.string()).default([]),
-    active: z.number().optional(),
-    children: z.array(z.lazy(() => pageComponentSchema)).default([]),
-  });
-
-export const pageComponentSchema: z.ZodType<PageComponent> = z.lazy(() =>
-  z.discriminatedUnion("type", [
-    announcementBarComponentSchema,
-    heroBannerComponentSchema,
-    valuePropsComponentSchema,
-    reviewsCarouselComponentSchema,
-    productGridComponentSchema,
-    productCarouselComponentSchema,
-    recommendationCarouselComponentSchema,
-    galleryComponentSchema,
-    lookbookComponentSchema,
-    contactFormComponentSchema,
-    newsletterSignupComponentSchema,
-    searchBarComponentSchema,
-    contactFormWithMapComponentSchema,
-    mapBlockComponentSchema,
-    storeLocatorBlockComponentSchema,
-    videoBlockComponentSchema,
-    faqBlockComponentSchema,
-    countdownTimerComponentSchema,
-    headerComponentSchema,
-    footerComponentSchema,
-    socialLinksComponentSchema,
-    socialFeedComponentSchema,
-    socialProofComponentSchema,
-    blogListingComponentSchema,
-    testimonialsComponentSchema,
-    pricingTableComponentSchema,
-    giftCardBlockComponentSchema,
-    popupModalComponentSchema,
-    testimonialSliderComponentSchema,
-    imageComponentSchema,
-    textComponentSchema,
-    customHtmlComponentSchema,
-    buttonComponentSchema,
-    sectionComponentSchema,
-    multiColumnComponentSchema,
-    tabsComponentSchema,
-  ])
-);
-
-export interface HistoryState {
-  past: PageComponent[][];
-  present: PageComponent[];
-  future: PageComponent[][];
-  gridCols: number;
-}
-
-export const historyStateSchema: z.ZodType<HistoryState> = z
-  .object({
-    past: z.array(z.array(pageComponentSchema)),
-    present: z.array(pageComponentSchema),
-    future: z.array(z.array(pageComponentSchema)),
-    gridCols: z.number().int().min(1).max(24).default(12),
-  })
-  .strict()
-  .default({ past: [], present: [], future: [], gridCols: 12 });
-
-export const pageSchema = z
-  .object({
-    id: z.string(),
-    slug: z.string(),
-    status: z.enum(["draft", "published"]),
-    components: z.array(pageComponentSchema).default([]),
-    seo: z.object({
-      title: z.record(localeSchema, z.string()),
-      description: z.record(localeSchema, z.string()).optional(),
-      image: z.record(localeSchema, z.string()).optional(),
-    }),
-    createdAt: z.string(),
-    updatedAt: z.string(),
-    createdBy: z.string(),
-    history: historyStateSchema.optional(),
-  })
-  .strict();
-
-export type Page = z.infer<typeof pageSchema>;
