--- packages/types/src/page/gallery.ts
@@ -1,1 +1,0 @@
-export { GalleryComponent, galleryComponentSchema } from "./base";
