--- packages/types/src/page/hero-banner.ts
@@ -1,1 +1,0 @@
-export { HeroBannerComponent, heroBannerComponentSchema } from "./base";
