+++ packages/types/src/page/index.ts
@@ -1,5 +1,6 @@
 export * from "./base";
-export * from "./hero-banner";
-export * from "./gallery";
-export * from "./value-props";
-export * from "./product-grid";
+export * from "./atoms";
+export * from "./molecules";
+export * from "./organisms";
+export * from "./layouts";
+export * from "./page";
