--- packages/types/src/page/product-grid.ts
@@ -1,1 +1,0 @@
-export { ProductGridComponent, productGridComponentSchema } from "./base";
