--- packages/types/src/page/value-props.ts
@@ -1,1 +1,0 @@
-export { ValuePropsComponent, valuePropsComponentSchema } from "./base";
